refactor: Modal 관련 컴포넌트를 공용 컴포넌트 디렉토리로 이동
