+++ components/AuthModal/AuthModal.tsx
@@ -2,8 +2,8 @@
 
 import CredentialsLogin from "components/AuthModal/CredentialsLogin";
 import SocialLogin from "components/AuthModal/SocialLogin";
-import Modal from "components/Modal/Modal";
-import ModalHeader from "components/Modal/ModalHeader";
+import Modal from "components/common/Modal/Modal";
+import ModalHeader from "components/common/Modal/ModalHeader";
 import { ModalComponentProps } from "types/modal";
 
 import styles from "./AuthModal.module.scss";
