+++ components/common/Modal/Modal.tsx
@@ -1,8 +1,8 @@
 "use client";
 
-import ModalBackground from "components/Modal/ModalBackground";
-import ModalContainer from "components/Modal/ModalContainer";
-import ModalPortal from "components/Modal/ModalPortal";
+import ModalBackground from "components/common/Modal/ModalBackground";
+import ModalContainer from "components/common/Modal/ModalContainer";
+import ModalPortal from "components/common/Modal/ModalPortal";
 import EventType from "types/event";
 import { ModalComponentProps } from "types/modal";
